requirements.txt
paddlepaddle-gpu
paddlex
